Historic Commit
Contains individually working (but yet to be tested together) code for drivetrain, LIDAR, and Line Follow. LIDAR should be all set in terms of returning a correct value - only note here is the sensor may need to be calibrated properly. For drivetrain, we'll need to work more with Matt and Jeff to determine proper settings. Line follow needs to be tested and tuned further, but it is functional.