Changes to Climber Code
This code automates the elevator-with-climber system. If the elevator is detected to be going down under the robot, the stilt will follow at the same speed, keeping the robot level automatically. This code also includes the option for drive team to manually move the climber - all controls in Robot.java.